+++ companionpages/companionpages/urls.py
@@ -10,6 +10,7 @@ from members.forms import MemberForm
 urlpatterns = patterns(
     '',
     url(r'^$', HomeView.as_view(), name='rmc_home'),
+    url(r'^api/v1/', include('api.urls')),
     url(r'^members/edit', 'profiles.views.edit_profile', {'form_class': MemberForm, }, name='rmc_edit'),
     url(r'^members/', include('profiles.urls')),
     # profiles/create/ named profiles_create_profile
@@ -20,15 +21,16 @@ urlpatterns = patterns(
     url(r'^contact/', ContactView.as_view(template_name='contact.html'),
         name='envelope-contact'),
     url(r'^faq/', FaqView.as_view(), name='rmc_faq'),
-    url(r'^news/', include('news.urls')),
     url(r'^admin/doc/', include('django.contrib.admindocs.urls')),
     url(r'^admin/', include(admin.site.urls)),
-    url(r'^companionpages/', include('supportingmaterials.urls')),
+    url(r'^compendia/', include('supportingmaterials.urls')),
 )
 
 urlpatterns += patterns(
     'django.contrib.flatpages.views',
-    url(r'^about/', 'flatpage', {'url': '/about/'}, name='rmc_about'),
-    url(r'^terms/', 'flatpage', {'url': '/terms/'}, name='rmc_terms'),
-    url(r'^privacy/', 'flatpage', {'url': '/privacy/'}, name='rmc_privacy'),
+    url(r'^about/', 'flatpage', {'url': '/about/'}, name='about'),
+    url(r'^terms/', 'flatpage', {'url': '/terms/'}, name='terms'),
+    url(r'^privacy/', 'flatpage', {'url': '/privacy/'}, name='privacy'),
+    url(r'^partners/', 'flatpage', {'url': '/partners/'}, name='partners'),
+    url(r'^developers/', 'flatpage', {'url': '/developers/'}, name='developers'),
 )