+++ HISTORY.rst
@@ -3,13 +3,20 @@
 History
 -------
 
+1.0.0-alpha (2013-10-24)
+++++++++++++++++++++++++
+
+* Simplified creation form
+* Added admin action to flip items from draft to active
+* Many template changes
+
 1.0.0-alpha (2013-10-05)
-++++++++++++++++
+++++++++++++++++++++++++
 
 * Many stylistic changes
 * Started using semantic versioning
 
 0.11.0 (2013-09-17)
-++++++++++++++++++
++++++++++++++++++++
 
 * First release with docs